Avoid type checking private_constants (side_eff) again during Qed (#4357).
Side effects are now an opaque data type, called private_con, you can
only obtain from safe_typing.  When add_constant is called on a
definition_entry that contains private constants, they are inlined in
the main proof term but not re-checked.
As a safety measure, the opaque data type contains a pointer to the
revstruct (an internal field of safe_env that changes every time a new
constant is added), and such pointer is compared with the current value
store in safe_env when the private_constant is inlined.  Only if the
comparison is successful then the private_constant is not re-checked,
else it is.  In short, we accept into the kernel private constant only
when they arrive in the very same order and on top of the very same env
they arrived when we fist check them.

Limitation: nothing is done for Defined, in this case we check
abstracted subproofs again.  There is no hard problem to solve, it is
just that the API is not there (something like
add_private_constant_as_global_constant is missing).

Also, private_constants produced by workers never pass the safety
measure (the revstruct pointer is an Ephemeron).  Sending back the
entire revstruct is possible but: 1. we lack a way to quickly compare
two revstructs, 2. it can be large.