[feedback] Remove `ErrorMsg` in favor of `Message Error`.
The ErrorMsg datatype was introduced to allow locations in messages,
however, it was redundant with error and used only in one place.

We remove it in favor of a more uniform treatment of messages with
location. This patch also removes the use of `Loc.ghost` in one place.

Lightly tested.