=> now a tactical
I had to port some hacks from ssr, because the TACTIC EXTEND api
requires tactics to start with a symbol.

Also, I uncovered a horrendous hack in the Ltac API. If a tactic takes
an argument of type ident, or a list, or a few other types, an implicit
Goal.enter is done, which changes the semantics of the tactic...

I don't know if => should do a Goal.enter or not. For example, if we
want:

case; cycle 1 => [ | ]

to do the intuitive thing (cycle should do the permutation), we need
either to change ssr's associativity:

(case; cycle 1) => [ | ] instead of case; (cycle 1 => [ | ]) which fails

or to make => really multi goal